und_flush_frame: don't overrun iovidx
since numiov is the number of iovec to be passed to writev,
it's one greater than the number in und->iov.  This is almost
certainly OK since und->iov is zeroed and there should be no
case where iovidx > NUMCELL-2, but it's sloppy.

Instead, correctly iterate up to und->iovidx+1 rather than numiov
when indexing in und->iov.