insert(): cursor-based overwrite rather than memmove when memory-limited
memmove()ing all the time is obviously not very efficient; but we'd prefer
not to have an additional index (or pointer) per bucket.  We can get a
decent pseudorandom replacement policy by tracking a single cursor modulo
2^32 and modding it to the appropriate range on each bucket overwrite.

This also points out that we really want to avoid single-entry buckets, so
allow a little bit of slack on the memory limit to ensure each bucket gets
at least two entries.